refactor build_memory_config
